feat(finance): add PDF export and scheduled reports with bug fixes
- Add OpenPDF dependency for PDF generation
- Implement PdfReportService: core PDF utility with proper BaseFont initialization
- Implement InvoicePdfService: PDF generation for legacy Invoice entities
- Implement MovePdfService: PDF generation for Odoo-style account.move invoices
- Implement FinancialReportPdfService: PDF for P&L, Balance Sheet, Trial Balance, General Ledger
- Implement ScheduledReportService: CRUD for scheduled report configurations
- Implement ReportScheduler: cron-based scheduler for automated report generation
- Implement EmailService: email delivery with optional JavaMailSender
- Add ScheduledReportController: REST API for managing scheduled reports
- Add ScheduledReport entity with JSONB report_params field
- Add V36 migration: scheduled_reports table
- Add V37 migration: fix journal_type column on journal_entries
- Add V38 migration: ensure CRM tables (leads, opportunities, pipeline_stages) exist
- Update ReportController: add PDF endpoints for all 4 financial reports
- Update InvoiceController: add PDF endpoint for invoice export
- Fix OpenPDF v2.0.3 font initialization (BaseFont.createFont instead of Font.HELVETICA)
- Fix Hibernate JSONB mapping with @JdbcTypeCode(SqlTypes.JSON)
- Fix CrmController ambiguous mapping collision with LeadController
- Fix InvoiceService and InvoiceDto compilation errors
- Fix SalesOrderService and SalesOrderController compilation errors
- Fix Flyway migrations V33-V35 idempotent (IF NOT EXISTS)
- Fix 13 test assertions: 403 FORBIDDEN → 401 UNAUTHORIZED for unauthenticated requests
- All 104 tests passing

+++ src/main/java/com/erp/sales/service/SalesOrderService.java
@@ -351,7 +351,7 @@ public class SalesOrderService {
             moveLine.setName(line.getProduct() != null ? line.getProduct().getName() : "Order Line");
             moveLine.setDebit(line.getLineTotal());
             moveLine.setCredit(BigDecimal.ZERO);
-            moveLine.setQuantity(line.getQuantity());
+            moveLine.setQuantity(BigDecimal.valueOf(line.getQuantity()));
             moveLines.add(moveLine);
         }
         move.setLines(moveLines);
@@ -370,7 +370,7 @@ public class SalesOrderService {
             seq++;
             Product product = productClient.getProductById(lineRequest.getProductId());
 
-            BigDecimal lineTotal = lineRequest.getUnitPrice().multiply(lineRequest.getQuantity());
+            BigDecimal lineTotal = lineRequest.getUnitPrice().multiply(BigDecimal.valueOf(lineRequest.getQuantity()));
             BigDecimal discount = lineRequest.getDiscount() != null ? lineRequest.getDiscount() : BigDecimal.ZERO;
 
             // Apply discount to line total
@@ -412,7 +412,7 @@ public class SalesOrderService {
             seq++;
             Product product = productClient.getProductById(lineRequest.getProductId());
 
-            BigDecimal lineTotal = lineRequest.getUnitPrice().multiply(lineRequest.getQuantity());
+            BigDecimal lineTotal = lineRequest.getUnitPrice().multiply(BigDecimal.valueOf(lineRequest.getQuantity()));
             BigDecimal discount = lineRequest.getDiscount() != null ? lineRequest.getDiscount() : BigDecimal.ZERO;
 
             BigDecimal discountedTotal = lineTotal;
